renomeia as variáveis para terem uma semântica melhor, muda os calculos de hora para o tipo DateTime. usa o tipo int na classe DadosIo

+++ Estudos/ConsoleApp1/Utils/DadosIo.cs
@@ -2,7 +2,7 @@
 {
     public class DadosIo
     {
-        public long SolicitarValorInteiroAoUsuario(string descricao, long? minValue = null, long? maxValue = null)
+        public int SolicitarValorInteiroAoUsuario(string descricao, long? minValue = null, long? maxValue = null)
         {
             var numeroDecimal = SolicitarValorDecimalAoUsuario(descricao, minValue, maxValue);
